server/requirements.txt
fastapi~=0.84.0
uvicorn~=0.23.2
sqlalchemy~=2.0.19
psycopg2-binary~=2.9.6
pydantic~=1.7.1
starlette~=0.19.1
jinja2~=3.1.2
python-multipart~=0.0.6
PyJWT~=2.8.0
passlib~=1.7.4
bcrypt~=4.1.3
asyncpg~=0.29.0